python_tools/pyproject.toml
[project]
name = "python-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "markitdown[all]>=0.1.3",
    "chardet>=5.2.0",
    "httpx>=0.27.0",
    "numpy>=2.0.0",
    "python-dotenv>=1.0.0",
    "gguf>=0.10.0",
]
